all tests work for numpy and torch backend (except for GeneralGaussian and sample ratSpn

+++ src/spflow/tensorly/sampling/spn/layers/cond_sum_layer.py
@@ -75,7 +75,7 @@ def sample(
 
         # single node id
         node_id = node_ids[0]
-        node_instance_ids = tl.tensor(sampling_ctx.instance_ids)[indices]
+        node_instance_ids = tl.tensor(sampling_ctx.instance_ids,dtype=int)[indices]
 
         sample(
             sum_layer.nodes[node_id],
